refactor(ui): folder structure - to classfy by file type - ui,hooks,utils - no pascal case for folder/file name
