+++ packages/ui/src/component/slider/ui/index.tsx
