+++ packages/ui/src/component/slider/ui/recipe.ts
