+++ packages/ui/src/component/tabs/ui/recipe.ts
