+++ packages/ui/src/stories/Button.stories.tsx
@@ -1,5 +1,5 @@
 import { expect, fn, userEvent, within } from "@storybook/test"
-import { Button } from "../component/Button"
+import { Button } from "../component/button/ui"
 import type { Meta, StoryObj } from "@storybook/react"
 export default {
   title: "Button",
